Pursuant to Issue #12, convert almost all the arrays to allocatable.
This did not have an appreciable effect on the runtime of MIM, under
-Ofast without -fcheck=all, on the examples in the test suite (at
100x100 resolution), however, it does seem to slow down the overall
execution time of the test suite, for some reason.

An issue: Apparently, allocatable arrays cannot be read by namelist
input, because the compiler generates the code for parsing namelists
at compile time, and for some reason needs to know how many elements
to expect.  This restriction is allegedly lifted in Fortran 2003, but
that doesn't seem like a great reason to change standards.

The practical consequence is that the per-layer parameters ah, g_vec,
and hmean will have to be read in some other way.