Swap loop nesting order of i and j in evaluate_dudt.
Why?  I profiled a reduced gravity run (500x500x1 by 502 time steps),
and got this:

  %   cumulative   self              self     total
 time   seconds   seconds    calls  ms/call  ms/call  name
 36.62     11.94    11.94                             MAIN__
 34.99     23.35    11.41      506    22.55    22.55  evaluate_dudt_
 11.47     27.09     3.74      506     7.39     7.39  evaluate_dvdt_

Why in the world was evaluate_dudt_ taking 3x as long as
evaluate_dvdt_, since they should be doing the same thing?  Turns out
the difference was that evaluate_dvdt_ was walking along in the
recommended loop nesting order, and evaluate_dudt_ had its i and j
iteration nesting swapped.

So I fixed it, and now the profile says this:

  %   cumulative   self              self     total
 time   seconds   seconds    calls  ms/call  ms/call  name
 48.01     11.81    11.81                             MAIN__
 15.81     15.70     3.89      506     7.69     7.69  evaluate_dudt_
 14.92     19.37     3.67      506     7.25     7.25  evaluate_dvdt_

evaluate_dudt_ accelerates by 3x, which reduces the total run time on
this example from 36s to 28.5s, an improvement of 21%.

Eyeballing the code, this should have no effect on the answers, since
I believe each iteration only writes to dudt(i,j,k), and doesn't read
the dudt array at all, so the results should be completely independent
of iteration order.

Also ran the test suite and it passed.