Pursuant to Issue #48, hack together basic hardware event counting for MIM runs.
This is presumably OS-specific, in that, on the advice of
http://www.bnikolic.co.uk/blog/hpc-howto-measure-flops.html, it uses
`perf stat`, which relies on a Linux API for measuring CPU events.  On
Ubuntu, `perf` can be obtained by installing the `linux-tools-generic`
package.

That weird number 530010 is a hex code for "floating point operation"
for my particular CPU, which I discovered through the somewhat annoying
process of building `perfmon2/libpfm4` from source, per the instructions
on http://askubuntu.com/questions/637144/how-to-install-perfmon2-on-ubuntu,
and then running the resulting `showevtinfo` and `check_events` programs,
as advised in http://www.bnikolic.co.uk/blog/hpc-prof-events.html.