Direct outputs indicating error conditions to standard error, instead of various log files.
Closes #45.

One caveat: Apparently, access to the standard error stream is
compiler-dependent in Fortran.  gfortran causes the compiled program
to read an environment variable, namely GFORTRAN_STDERR_UNIT, to
determine how to expose the standard error stream.

This commit sets that variable in the Python controller code, but that
strategy will not work on a different Fortran compiler without
porting.  I chose `17` as the unit number so it will be relatively
easy to search for.