Try adding an open mpi runtime package to the Travis config.
Groveling the Ubuntu packages database at packages.ubuntu.com suggests
that `mpirun` is not in the libopenmpi-dev package, which would
explain Travis's inability to execute the test suite.  It's fairly
standard in Debian-based distros such as Ubuntu to split the
development files (libfoo) from the runtime support (openmpi-bin, in
this case), because there are purposes that need one but not the
other.  (e.g., just running a pre-compiled MPI application does not
need libopenmpi-dev).