Add ub, vb, etanew, and etastar to outputs
When debug_level >= 4 the model now outputs the internal variables for the barotropic correction.
The python module also recognises them and can read them in.

This commit also modifies interpret_raw_files to make use of the dimensionality pun discussed in #116. All fields now have three axes. For 2D fields, the z axis has length one. This will allow the output preservation tests to use the same code snippet for plotting, regardless of the field that triggers the assertion error.