Add example config file and docs, remove old one
A version of the new aronnax.conf file has been added to the base directory of the repo. This config file includes all of the paramters and inputs that can be specified as well as comments describing them. This partially completes #103. The documenation page `running_aronnax` ingests this config file and displays it, along with its comments and some additional descriptive text.