implement boundary conditions in A matrix
The boundary conditions were previously implemented in a very basic way assuming that the boudnary was the edge of the domain. This new code uses the boundary masks to apply boundary conditions everywhere that wetmask tansitions from one to zero. This modification causes the test suite to fail, since the boundary conditions are calculated slightly differently.

Implementing these boundary conditions finalises the implementation of periodic BCs in all versions of the model, and closes #15.