periodic BC tests
Add two tests to the test suite. These begin with a bump in layer thickness or free surface, and periodic BC. The waves travel away from the initial disturbance and cross each of the boundaries in turn. Plotting the outputs of the test suite (at a higher resolution would be nicer) allows one to watch the waves crossing the boundaries without altering speed or direction - as expected with periodic BCs. The n-layer version is set to use the Hypre solver, which extends the coverage of the test suite, but means there is no periodic BC test of the internal SOR solver.