remove allocate commands from model_run
The allocate commands were introduced in #72 to prevent the model from placing large arrays on the stack when using the -Ofast compiler flag. When the arrays were too large for the stack the model would segfault. As discussed in #73, using the -fno-stack-arrays compiler flag achieves the same outcome, does not affect performance, and allows for the removal of the allocate commands from the source code, thereby simplifying the code and improving readability.

This closes #73.