convert 2d and 3d u and v generators
The 2d and 3d versions of the u and v point generator functions have been combined into single functions that take a list of arguments and assert that the list has the appropriate length for the field being created. The test suite and reproductions have been updated to run with this new configuration. This closes #116