optionally make n-layer double gyre run go longer
If this simulation doesn't have to run everytime on Travis, it can take longer. Hence, the script now contains a line to pass the same nTimeSteps and dumpFreq options as are used in the reduced gravity version. The line is commented out by default, so that users won't get hit with a 20 min run when they try the examples script.