teach model to read checkpoint files
if niter0 is not zero, then the model looks in 'checkpoints' folder for files to restart the simulation.

This required correcting the checkpoint arrays (the halos were being excluded and only one time level of tendencies was being saved).
The wind_mag_time_series is assumed to start at the same time step as the simulation.