block output of averaged fields at first timestep
As discussed in #107 the average fields dumped after the first timestep are not correct. Even if they were correct, they wouldn't be of much use, since averaging for a single timestep doesn't make sense.

This commit blocks the output of these fields, and removes the equivalent blesssed output. This closes #107.