Make test suite compare all output files
Previous behaviour was to stop at the first file that failed the test. It now compares all files, and if the relative error is above the tolerance value, then it plots the blessed output, the current output, and the difference. Once all files have been compared the assertion error is raised and the test fails. This produces many image files and shows how the difference evolves over time.

Closes #136