Allow wind forcing to act on other layers
This caused the dudt and dvdt subroutines to become very large. They were therefore refactored, so that each contribution to the tendency is calculated in its own subroutine.

If `wind_depth` is set to zero, then the wind forcing will act only on the upper layer, no matter how thin it gets (preserving the previous behaviour).