Change enforcement of minimum layer thickness
Previous behaviour was to force layers to be thicker than hmin. The new behaviour is to exponentially increase thickness diffusivity as h approaches hmin, as suggested in a comment on #26. This produces very thin flat layers that allow the model to continue running, and don't affect volume conservation as badly.